refactor: some token-detail components
