+++ packages/web/src/views/token-detail/components/trending-crypto-card-list/TrendingCryptoCardList.styles.ts
