+++ packages/web/src/views/token-detail/components/trending-crypto-card-list/trending-crypto-card/TrendingCryptoCard.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
 import TrendingCryptoCard from "./TrendingCryptoCard";
-import { trendingCryptoListInit } from "@containers/trending-crypto-card-list-container/TrendingCryptoCardListContainer";
+import { trendingCryptoListInit } from "@views/token-detail/containers/trending-crypto-card-list-container/TrendingCryptoCardListContainer";
 
 export default {
   title: "token/TrendingCryptoCard",
