+++ packages/web/src/views/token-detail/components/trending-crypto-card-list/trending-crypto-card/TrendingCryptoCard.styles.ts
